# app/static/swagger.yaml
swagger: "2.0"
info:
  title: Mechanic Shop API
  version: "1.0"
  description: API documentation for Mechanic Shop
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths: 
...
  /service_tickets:
    post:
      tags:
        - Service Tickets
      summary: "Endpoint to create/register a new service ticket."
      description: "Endpoint creates a new service ticket in our library system."
      parameters:
        - in: body
          name: "body"
          description: "Information required to create a new service ticket."
          required: true
          schema:
            $ref: "#/definitions/CreateServiceTicketPayload"
      responses:
        '201':
          description: "Service ticket created successfully"
          schema:
            $ref: "#/definitions/CreateServiceTicketResponse"
          examples:
            application/json:
              id: 1
              description: "Windshield Wipers installation"
              customer_id: "1"
    get:
      tags:
        - Service Tickets
      summary: "Returns all service tickets"
      description: "Endpoint to retrieve a list of all service tickets"
      responses:
        '200':
          description: "Endpoint to retrieve a list of all service tickets."
          schema:
            $ref: "#/definitions/AllServiceTickets"
  /service_tickets/{id}:
    put: 
      tags:
        - Service Tickets
      summary: "Endpoint to update service_tickets."
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
        - in: body
          name: body
          description: Information required to update service ticket item.
          required: true
          schema:
            $ref: "#/definitions/UpdateServiceTicketsPayload"
      responses:
        '200':
          description: "Successfully Updated Service Ticket item"
          schema:
            $ref: "#/definitions/UpdateServiceTicketsResponse"
          examples:
            application/json:
              id: 1
              description: "Windshield Wipers replacement"
              customer_id: "2"

    delete:
     tags:
      - Service Tickets
     summary: "Endpoint to delete a service ticket."
     description: "Endpoint to delete a service ticket."
     security: 
      - bearerAuth: []
     responses:
       '200':
         description: "Successfully deleted service ticket item"
         schema:
           $ref: "#/definitions/DeleteServiceTicketResponse"
         examples:
           application/json:
             message: "Successfully deleted service ticket item <service_ticket_id>."
definitions: 
...
  CreateServiceTicketPayload:
    type: "object"
    properties:
      id:
        type: "string"
      description:
        type: "string"
      customer_id:
        type: "string"
    required:
      - id
      - description
      - customer_id
  CreateServiceTicketResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      description:
        type: "string"
      customer_id:
        type: "string"
  AllServiceTickets:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
        description:
          type: "string"
        customer_id:
          type: "string"
...
  DeleteServiceTicketResponse:
   type: "object"
   properties:
    message:
     type: "string"